Restructuring of services and commands:
-- services are responsible for working with their entities
-- commands is class, that contains all commands for user and unites all services (adding new logic, if services intersect)

+++ src/main/java/com/andersentask/bookshop/console/Commands.java
@@ -17,25 +17,19 @@ public class Commands {
 
 
     /**
-     * Set status of book, if bookStatus of param is opposite to the current status
-     * This method also deletes all requests on book, if status is changed to available
+     * Set status of book
+     * This method also deletes all requests on book,
+     * if it's status changed from out_of_stock to available
      *
      * @param id         the id of the book and should be got from user
      * @param bookStatus bookStatus of the book and got from user
      */
     public void setStatusToBookAndDeleteCorrespondingRequests(Long id, BookStatus bookStatus) {
-        Optional<Book> optionalBook = appContextConfig.getBookService().getBookById(id);
+        Optional<Book> optionalBook = appContextConfig.getBookService().setStatusToBook(id,bookStatus);
         if (optionalBook.isPresent()) {
             Book book = optionalBook.get();
-
-            switch (bookStatus) {
-                case OUT_OF_STOCK -> book.setStatus(BookStatus.AVAILABLE);
-                case AVAILABLE -> {
-                    if (book.getStatus().equals(BookStatus.OUT_OF_STOCK)) {
-                        book.setStatus(BookStatus.AVAILABLE);
-                        appContextConfig.getRequestService().deleteRequest(book);
-                    }
-                }
+            if (book.getStatus().equals(BookStatus.OUT_OF_STOCK) && bookStatus.equals(BookStatus.AVAILABLE)) {
+                appContextConfig.getRequestService().deleteRequest(book);
             }
         }
     }
@@ -47,53 +41,42 @@ public class Commands {
      * @param bookSort can be name, price or status and should be got from user
      * @return new list of books, sorted by entered param
      */
-    public List<Book> getSortedBooks(BookSort bookSort) {
-        List<Book> books = appContextConfig.getBookService().getAllBooks();
-        List<Book> booksToReturn = new ArrayList<>();
-        switch (bookSort) {
-            case NAME -> booksToReturn = books.stream()
-                    .sorted(Comparator.comparing(Book::getName))
-                    .toList();
-            case PRICE -> booksToReturn = books.stream()
-                    .sorted(Comparator.comparing(Book::getPrice))
-                    .toList();
-            case STATUS -> booksToReturn = books.stream()
-                    .sorted(Comparator.comparing(x -> x.getStatus().ordinal()))
-                    .toList();
-        }
-        return booksToReturn;
+    public List<Book> getBooks(BookSort bookSort) {
+         return appContextConfig.getBookService().getSortedBooks(bookSort);
     }
 
 
     /**
      * Create and save request from a book (doesn't matter, what status)
      *
-     * @param book book object and should be got from user
+     * @param id book object and should be got from user
      */
-    public void createRequest(Book book) {
-        Request request = appContextConfig.getEntityFactory().buildRequest(book);
-        appContextConfig.getRequestService().saveRequest(request);
+    public void createRequest(Long id) {
+        Optional<Book> optionalBook = appContextConfig.getBookService().getBookById(id);
+        if (optionalBook.isPresent()) {
+            Book book = optionalBook.get();
+            Request request = appContextConfig.getEntityFactory().buildRequest(book);
+            appContextConfig.getRequestService().saveRequest(request);
+        }
     }
 
 
     /**
-     * Create and save order from a list of book
+     * Create and save order from a list of id of books
      * If book is out_of_stock, also created ans saves request
+     * if any id is invalid, the order and request are not created
      *
-     * @param books list of books and should be got from user
+     * @param ids list of id of books and should be got from user
      */
-    //toDO: change books
-    public void createOrder(List<Book> books) {
-        Order order = appContextConfig.getEntityFactory().buildOrder(books);
-        appContextConfig.getOrderService().saveOrder(order);
+    public void createOrder(List<Long> ids) {
+        List<Book> booksToOrder = appContextConfig.getBookService().getBooksByIds(ids);
 
-        List<Book> booksOutOfStock = books.stream()
-                .filter(x -> x.getStatus().equals(BookStatus.OUT_OF_STOCK))
-                .toList();
-        if (!booksOutOfStock.isEmpty()) {
-            for (Book book : booksOutOfStock) {
-                appContextConfig.getEntityFactory().buildRequest(book);
-            }
+        if (booksToOrder.size() == ids.size()) {
+            Order order = appContextConfig.getEntityFactory().buildOrder(booksToOrder);
+            appContextConfig.getOrderService().saveOrder(order);
+
+            List<Book> booksToRequest = appContextConfig.getBookService().getBooksOutOfStock(booksToOrder);
+            booksToRequest.forEach(x -> appContextConfig.getEntityFactory().buildRequest(x));
         }
     }
 
@@ -111,25 +94,13 @@ public class Commands {
         if (optionalOrder.isPresent()) {
             Order order = optionalOrder.get();
 
-            switch (orderStatus) {
-                case COMPLETED -> {
-                    boolean allBooksAreAvailable = order.getBooksInOrder().stream()
-                            .allMatch(x -> x.getStatus().equals(BookStatus.AVAILABLE));
-                    if (allBooksAreAvailable) {
-                        order.setOrderStatus(OrderStatus.COMPLETED);
-                        order.setTimeOfCompletingOrder(LocalDateTime.now());
-                    }
+            if (orderStatus.equals(OrderStatus.COMPLETED)) {
+                boolean allBooksAreAvailable = appContextConfig.getBookService().allBooksAreAvailable(order.getBooksInOrder());
+                if (allBooksAreAvailable){
+                    appContextConfig.getOrderService().changeStatusOfOrder(id,orderStatus);
                 }
-                case CANCELED -> {
-                    if (order.getOrderStatus().equals(OrderStatus.IN_PROCESS)) {
-                        order.setOrderStatus(OrderStatus.CANCELED);
-                    }
-                }
-                case IN_PROCESS -> {
-                    if (order.getOrderStatus().equals(OrderStatus.CANCELED)) {
-                        order.setOrderStatus(OrderStatus.IN_PROCESS);
-                    }
-                }
+            } else {
+                appContextConfig.getOrderService().changeStatusOfOrder(id, orderStatus);
             }
         }
     }
@@ -141,21 +112,8 @@ public class Commands {
      * @param orderSort can be cost, completion_date or status and should be got from user
      * @return new list of orders, sorted by entered param
      */
-    public List<Order> getSortedOrders(OrderSort orderSort) {
-        List<Order> orders = appContextConfig.getOrderService().getAllOrders();
-        List<Order> ordersToReturn = new ArrayList<>();
-        switch (orderSort) {
-            case COST -> ordersToReturn = orders.stream()
-                    .sorted(Comparator.comparing(Order::getOrderCost))
-                    .toList();
-            case COMPLETION_DATE -> ordersToReturn = orders.stream()
-                    .sorted(Comparator.comparing(Order::getTimeOfCompletingOrder).reversed())
-                    .toList();
-            case STATUS -> ordersToReturn = orders.stream()
-                    .sorted(Comparator.comparing(x -> x.getOrderStatus().ordinal()))
-                    .toList();
-        }
-        return ordersToReturn;
+    public List<Order> getOrders(OrderSort orderSort) {
+        return appContextConfig.getOrderService().getSortedOrders(orderSort);
     }
 
     /**
@@ -165,22 +123,17 @@ public class Commands {
      * @return Long number of the request on the precise book
      */
     public Long getNumberOfRequestsOnBook(Long id) {
-        return appContextConfig.getRequestService().getAllRequests().stream()
-                .map(Request::getBook)
-                .filter(x -> x.getId().equals(id))
-                .count();
+        return appContextConfig.getRequestService().getNumberOfRequestsOnBook(id);
     }
 
     /**
-     * return books and the number of requests on these books in descending order
+     * return id of books and the number of requests on these books in descending order
      *
-     * @return Map<String, Long> name of the book and number of requests on this book
+     * @return Map<Long, Long> id of the book and number of requests on this book
      */
-    private Map<Long, Long> getBooksAndNumberOfRequests() {
+    public Map<Long, Long> getBooksAndNumberOfRequests() {
         Map<Long, Long> map = new HashMap<>();
-        List<Book> books = appContextConfig.getRequestService().getAllRequests().stream()
-                .map(Request::getBook)
-                .toList();
+        List<Book> books = appContextConfig.getRequestService().getAllBooksFromRequests();
         List<Book> booksDistinctSortedByNumber = books.stream()
                 .distinct()
                 .sorted(Comparator.comparing(x -> getNumberOfRequestsOnBook(x.getId()), Comparator.reverseOrder()))
@@ -200,21 +153,17 @@ public class Commands {
      * @return double income for the chosen period
      */
     public double getIncomeForPeriod(LocalDateTime startOfPeriod, LocalDateTime endOfPeriod) {
-        return appContextConfig.getOrderService().getAllOrders().stream()
-                .filter(x -> x.getOrderStatus().equals(OrderStatus.COMPLETED))
-                .filter(x -> x.getTimeOfCompletingOrder().isAfter(startOfPeriod) &&
-                        x.getTimeOfCompletingOrder().isBefore(endOfPeriod))
-                .map(Order::getOrderCost)
-                .reduce(0D, Double::sum);
+        return appContextConfig.getOrderService().getIncomeForPeriod(startOfPeriod,endOfPeriod);
     }
 
     /**
      * return all books from the order
      *
+     * @param id id of the order and should be got from user
      * @return List<BooK> list of books
      */
-    public List<Book> getAllBooksFromOrder(Order order) {
-        return order.getBooksInOrder().stream().toList();
+    public List<Book> getAllBooksFromOrder(Long id) {
+        return appContextConfig.getOrderService().getAllBooksFromOrder(id);
     }
